Discipline DONE

+++ Academy/Models/Discipline.cs
@@ -6,7 +6,7 @@ namespace Academy.Models
     {
         [Key]
         [Required(ErrorMessage = "ID is required")]
-        public byte discipline_id { get; set; }
+        public short discipline_id { get; set; }
 
         [Required(ErrorMessage = "Discipline name is required")]
         [StringLength(100, MinimumLength = 2, ErrorMessage = "Name must be between 2 and 100 characters")]
